--- src/transactionsPage/transactionsPagePagination/index.ts
@@ -1,1 +1,0 @@
-export * from "./transactionsPagePagination";
